add python-gnupg dependency.
Also update requirements-{dev,prod,test-prod}.txt with python-gnupg and
the dephash output.  Since I'm not using any of those, it might be time
to get rid of them.  I could see them being useful to someone, though,
as long as they're kept up to date.  I'll punt on that decision for now.