add `scriptworker.client.validate_artifact_url`.
Scriptworker and its various scripts will be given various urls to
download artifacts from.  This is beneficial, as long as we can trust
those urls.  We have some level of trust through TaskCluster scopes, but
those are not foolproof.  We need to trust but verify that the links are
not going to evil.com.

Each scriptworker instance and client script can define a whitelist of
url schemes, netlocs, and path regexes.  The regexes can define how to
read the taskId of the link, which we can validate against a whitelist
(most likely the task's parent task ids).  The regexes, if defined, must
also define a filepath, which will be returned on successful url
validation.  If the regexes are not defined, the full path is the
filepath.

On failure, validate_artifact_url will throw a
`ScriptWorkerTaskException` with `malformed-payload`.