write task logs directly to task_log_dir
Originally, we wrote the task logs into log_dir along with the worker
logs, and then copied them into the artifact_dir before uploading.  This
is inefficient and doesn't allow for users to specify where the logs are
uploaded to (it hardcoded public/logs/).  Instead, let's define a
task_log_dir, and if it's a subdir of artifact_dir, the task logs get
uploaded as expected.  Then we can explicitly specify public/ or private/
inside of the task_log_dir config itself.