add 'decision' info to each cot graph dependency.
We're going to have to match each task to its decision task; this is
complicated by the fact that docker-image tasks have a different
decision task than the other tasks in the graph.  We could look up each
decision task from the task definition each time, but this is faster.

To do this I had to sort the decision task first in the
`build_task_dependencies` loop, and I turned the `dependencies` dict
into a nested dict.

This assumes that each Chain of Trust graph is going to have at least
one decision task.  Per #tcmigration this is a good assumption.