refactor validate_artifact_url with valid_artifact_rules
Previously, we had a list of valid schemes, a list of valid netlocs,
and a list of valid paths.  This was fine with a single entry in each,
but if we want to allow for two different sites, the paths will be valid
on either netloc, etc.  Splitting it into a single list of dictionaries
allows us to be more granular about this.