stricter gpg homedir lockfile usage.
Previously, we were creating the gpg lockfile before recreating the
worker gpg homedirs, but after recreating ~/.gnupg to verify the git
repo commits.  This meant we were recreating ~/.gnupg every five
minutes, and could have resulted in a long-running process overrunning
the next, especially if we ever shorten the frequency.

Now we create the lockfile before doing any of it, and clean up after
we're done.  I plan to file a bug about nagios checks for the age of the
lockfile and the various scriptworker logs, so we catch if any of these
take significantly longer than expected.