allow for beetmover and balrog restricted scopes.
When adding the beetmover and balrog restricted scopes, I noticed that
we redefine the nightly and release trees several times.  To avoid this,
I split this into two dicts: a scope-to-level dict, and a level-to-trees
dict.  I also am indexing by `cot_product`, because these tree lists
may likely be different for addons or servo or whatnot.