nuke and rename gpg base homedir
In the worker loop, we were using `overwrite_gpg_home`, which is
designed to overwrite a gpg homedir with a new one's contents.  However,
we were calling it against the parent directory of 3 gpg homedirs,
rather than individually for each gpg homedir.  This is why we broke
previously, and why we're still breaking when we rebuild gpg homedirs.

When we create the backup (gpg.old), we
  `rm('gpg.old'); os.rename('gpg', 'gpg.old')`
This nukes and moves the entire directory tree.  This is an easy fix to
apply to the new `gpg.tmp` as well.