Delete cancelled or claimed tasks from azure queue.
While testing the nightly migration on the `date` branch, we noticed
scriptworker was slogging through a lot of tasks that were "already
claimed by another worker".  These turned out to be cancelled tasks, and
they stayed in the Azure queue because we didn't delete them.

This patch deletes those tasks from the Azure queue on status code 409.
It also adds retries and catches in `find_task`, making it more robust.
This is already running on beetmoverworker-1, and we've cleared out a
backlog of cancelled tasks from the queue because of it.

This patch fixes #55.