Restrict googleplay scopes.
Also rename the current `release` and `nightly` branch aliases to
`all-release-branches` and `all-nightly-branches`, because they are a
superset of the repos for those releases.  Restricting `release`,
`beta`, and `nightly` to the actual repos used for those release trains
will let us avoid pushing the wrong apk to the wrong product (though the
product name helps avoid that as well).