allow for a subset of task dependencies in cot fuzzy matching.
Currently, SOP for release promotion is to cancel breakpoint tasks when
we're ready to push live, and then the downstream tasks can trigger.

Pushapk is scheduled as part of the relpro graph 1, with a breakpoint
task dependency.  However, we currently have to schedule a graph 2 when
we pass the expiry point; this is problematic, as graph 2 isn't
currently CoT enabled.

Fuzzy matching allows us to retrigger a task, and it allows for changed
timestamps and taskIds.  However, until now, the dependency list
had to stay the same.  This was problematic for any task with an expired
breakpoint task dependency.

This change allows us to retrigger the pushapk (or any other) task, with
updated timestamps and a *subset* of dependencies, allowing us to bypass
the need for a CoT-enabled graph 2.  We still want that, but in its own
timeframe.