add a log message to claimWork
Previously, we had polling logs in worker.log, so we knew that the
worker was running. With claimWork, we're silent until we claim a task,
which can be many hours in a row. With nagios monitoring our worker.log
age, we should output something, plus it's helpful to know that the
worker is running.