don't run docker-worker tests on the chain object
`verify_cot` breaks when run against a docker-worker task, because we
try to verify chain.cot for docker-worker specific info. Because the
chain object is assumed to be the currently running task, we don't
create a cot object for it.

I ran into this while testing

```
verify_cot --task-type l10n --cleanup -- Tz_iPxtTRqCLBg64rHK8uQ
```