allow for unsigned cot artifacts when verify_cot_signature is False
This python-gnupg behavior seems to have changed, or my memory is
faulty. I remember `gpg.decrypt` returning an unsigned body when the
input was signed, and returning the original string when the input
wasn't signed. Now we're dying on `verify_chain_of_trust: True`,
`verify_cot_signature: False`, and an unsigned cot artifact.

Get around this by explicitly returning the original string if
`verify_cot_signature` is `False`, and if `gpg.decrypt` returns an empty
string.