RELENG-830 - Replace the 2022 docker-worker CoT key
When I generated the CoT ed25519 keypair for docker-worker CoT, I had
used the Python script in braindump rather than the JS one found in
Taskcluster (because I had it readily available and I assumed they would
be the same).

But it turns out the Python script generates a key 32 character key,
while the JS script generates a 64 character key. Docker worker throws
an error when attempting to use the former.

To fix this, we had to generate a new keypair, thus this patch.