Bug 1869123 - Allow comm-esr115 to use the "balrog:server:release" restricted scope.
Fixes a regression from PR #624.

This is a temporary situation while Thunderbird monthly releases are in development.
Until comm-esr115 is migrated to use a (new) "esr" update channel, it needs to
use the "release" channel.

Details why comm-esr115 was able to use "release" prior to now are in the bug.