fix: ensure pull request head sha is set correctly
At the moment scriptworker assumes that the sha of the current tip of the PR is correct for all tasks upstream of anything being verified. This is incorrect if a task being verified depends on a cached task (whether from an earlier version of the same PR, or another one altogether).