Bug 1975771 - Don't spam logs if we continue receiving SIGUSR1
Sometimes the worker takes so long to start that it's already been told
to shut off before it sets its signal handlers which leads to the worker
continuing as if everything's going well. To combat that, we want
to send the USR1 signal in a loop instead of assuming that the worker
has already received it. Receiving said signal is idempotent so it
shouldn't be an issue but we don't want it to spam the logs if
possible